ArticleService/requirements.txt
astroid @ file:///Users/runner/miniforge3/conda-bld/astroid_1684150134139/work
beautifulsoup4==4.12.2
CacheControl==0.13.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
coverage==7.2.7
cryptography==41.0.1
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1666603105584/work
firebase-admin==6.1.0
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
isort @ file:///home/conda/feedstock_root/build_artifacts/isort_1675033873689/work
lazy-object-proxy @ file:///Users/runner/miniforge3/conda-bld/lazy-object-proxy_1672877875800/work
mccabe @ file:///home/conda/feedstock_root/build_artifacts/mccabe_1643049622439/work
msgpack==1.0.5
packaging==23.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1683850015520/work
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pylint @ file:///home/conda/feedstock_root/build_artifacts/pylint_1683397482155/work
pyparsing==3.0.9
pytest==7.3.1
pytest-cov==4.1.0
requests==2.31.0
requests-mock==1.10.0
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1644342247877/work
tomlkit @ file:///home/conda/feedstock_root/build_artifacts/tomlkit_1683028165096/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685704949284/work
uritemplate==4.1.1
urllib3==1.26.16
wrapt @ file:///Users/runner/miniforge3/conda-bld/wrapt_1677485661943/work
